Move misplaced analyzer tests to analyzer.test.ts
Move 6 analyzer-only tests from incorrect files to analyzer.test.ts:
- 3 valueBranches tests from jsx-to-ir.test.ts
- 2 isExported tests from adapter-output.test.ts
- 1 let declaration capture test from client-js-generation.test.ts
