+++ packages/jsx/src/__tests__/analyzer.test.ts
@@ -94,4 +94,135 @@ describe('analyzeComponent', () => {
     // Should NOT collect variables from nested function declaration
     expect(ctx.localConstants.some((c) => c.name === 'hash')).toBe(false)
   })
+
+  test('ternary constant has valueBranches', () => {
+    const source = `
+      'use client'
+      import { createSignal } from '@barefootjs/dom'
+
+      export function Demo() {
+        const [active, setActive] = createSignal(false)
+        const cls = active() ? 'a b' : 'c d'
+        return <div className={cls}></div>
+      }
+    `
+
+    const ctx = analyzeComponent(source, 'test.tsx')
+    const cls = ctx.localConstants.find(c => c.name === 'cls')
+    expect(cls).toBeDefined()
+    expect(cls!.valueBranches).toEqual(["'a b'", "'c d'"])
+  })
+
+  test('nested ternary constant has flattened valueBranches', () => {
+    const source = `
+      'use client'
+      import { createSignal } from '@barefootjs/dom'
+
+      export function Demo() {
+        const [state, setState] = createSignal(0)
+        const cls = state() === 0 ? 'a' : state() === 1 ? 'b' : 'c'
+        return <div className={cls}></div>
+      }
+    `
+
+    const ctx = analyzeComponent(source, 'test.tsx')
+    const cls = ctx.localConstants.find(c => c.name === 'cls')
+    expect(cls).toBeDefined()
+    expect(cls!.valueBranches).toEqual(["'a'", "'b'", "'c'"])
+  })
+
+  test('non-ternary constant has no valueBranches', () => {
+    const source = `
+      'use client'
+
+      export function Demo() {
+        const cls = 'hello world'
+        return <div className={cls}></div>
+      }
+    `
+
+    const ctx = analyzeComponent(source, 'test.tsx')
+    const cls = ctx.localConstants.find(c => c.name === 'cls')
+    expect(cls).toBeDefined()
+    expect(cls!.valueBranches).toBeUndefined()
+  })
+
+  test('export { X } named export syntax sets isExported', () => {
+    const source = `
+      'use client'
+      import { createSignal } from '@barefootjs/dom'
+
+      const MY_CONST = 42
+
+      export { MY_CONST }
+
+      export function Widget() {
+        const [val, setVal] = createSignal(0)
+        return <div>{MY_CONST}</div>
+      }
+    `
+    const ctx = analyzeComponent(source, 'Widget.tsx')
+    const constInfo = ctx.localConstants.find(c => c.name === 'MY_CONST')
+    expect(constInfo).toBeDefined()
+    expect(constInfo!.isExported).toBe(true)
+  })
+
+  test('isExported flag for export const, internal const, and export let', () => {
+    const source = `
+      'use client'
+      import { createSignal } from '@barefootjs/dom'
+
+      export const EXPORTED_A = 'aaa'
+      const INTERNAL_B = 'bbb'
+      export let EXPORTED_C = 100
+
+      export function MyComponent() {
+        const [val, setVal] = createSignal(0)
+        return <div />
+      }
+    `
+    const ctx = analyzeComponent(source, 'Test.tsx')
+
+    const a = ctx.localConstants.find(c => c.name === 'EXPORTED_A')
+    expect(a).toBeDefined()
+    expect(a!.isExported).toBe(true)
+    expect(a!.declarationKind).toBe('const')
+
+    const b = ctx.localConstants.find(c => c.name === 'INTERNAL_B')
+    expect(b).toBeDefined()
+    expect(b!.isExported).toBeFalsy()
+
+    const c = ctx.localConstants.find(c => c.name === 'EXPORTED_C')
+    expect(c).toBeDefined()
+    expect(c!.isExported).toBe(true)
+    expect(c!.declarationKind).toBe('let')
+  })
+
+  test('let without initializer is captured', () => {
+    const source = `
+      'use client'
+      import { createSignal, createEffect, onCleanup } from '@barefootjs/dom'
+
+      type ApiType = { scrollPrev: () => void }
+
+      export function Carousel() {
+        let emblaApi: ApiType | undefined
+        const [canScrollPrev, setCanScrollPrev] = createSignal(false)
+
+        createEffect(() => {
+          if (emblaApi) {
+            setCanScrollPrev(true)
+          }
+        })
+
+        return <div>{canScrollPrev() ? 'yes' : 'no'}</div>
+      }
+    `
+
+    const ctx = analyzeComponent(source, 'Carousel.tsx')
+    const letConstant = ctx.localConstants.find(c => c.name === 'emblaApi')
+    expect(letConstant).toBeDefined()
+    expect(letConstant!.declarationKind).toBe('let')
+    expect(letConstant!.value).toBeUndefined()
+  })
 })
