+++ packages/jsx/src/__tests__/jsx-to-ir.test.ts
@@ -110,56 +110,4 @@ describe('jsxToIR', () => {
       }
     }
   })
-
-  test('ternary constant has valueBranches', () => {
-    const source = `
-      'use client'
-      import { createSignal } from '@barefootjs/dom'
-
-      export function Demo() {
-        const [active, setActive] = createSignal(false)
-        const cls = active() ? 'a b' : 'c d'
-        return <div className={cls}></div>
-      }
-    `
-
-    const ctx = analyzeComponent(source, 'test.tsx')
-    const cls = ctx.localConstants.find(c => c.name === 'cls')
-    expect(cls).toBeDefined()
-    expect(cls!.valueBranches).toEqual(["'a b'", "'c d'"])
-  })
-
-  test('nested ternary constant has flattened valueBranches', () => {
-    const source = `
-      'use client'
-      import { createSignal } from '@barefootjs/dom'
-
-      export function Demo() {
-        const [state, setState] = createSignal(0)
-        const cls = state() === 0 ? 'a' : state() === 1 ? 'b' : 'c'
-        return <div className={cls}></div>
-      }
-    `
-
-    const ctx = analyzeComponent(source, 'test.tsx')
-    const cls = ctx.localConstants.find(c => c.name === 'cls')
-    expect(cls).toBeDefined()
-    expect(cls!.valueBranches).toEqual(["'a'", "'b'", "'c'"])
-  })
-
-  test('non-ternary constant has no valueBranches', () => {
-    const source = `
-      'use client'
-
-      export function Demo() {
-        const cls = 'hello world'
-        return <div className={cls}></div>
-      }
-    `
-
-    const ctx = analyzeComponent(source, 'test.tsx')
-    const cls = ctx.localConstants.find(c => c.name === 'cls')
-    expect(cls).toBeDefined()
-    expect(cls!.valueBranches).toBeUndefined()
-  })
 })
